+++ lib/departure/null_logger.rb
@@ -1,0 +1,15 @@
+module Departure
+  class NullLogger
+    def say(_message, _subitem = false)
+      # noop
+    end
+
+    def write(_text)
+      # noop
+    end
+
+    def write_no_newline(_text)
+      # noop
+    end
+  end
+end